% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.obs.imp.R
\name{compare.obs.imp}
\alias{compare.obs.imp}
\title{Get summary statistics for the observed and imputed parts of the data}
\usage{
compare.obs.imp(orig, imp)
}
\arguments{
\item{orig}{data frame of original (i.e. incomplete) data}

\item{imp}{list of length \code{m} of imputed data sets}
}
\description{
Get summary statistics for the observed and imputed parts of the data
}
\author{
Kristian Kleinke
}
