% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.percent.count.R
\name{compare.percent.count}
\alias{compare.percent.count}
\title{Compare percentages of observed and imputed counts}
\usage{
compare.percent.count(orig, imp, counts)
}
\arguments{
\item{orig}{data frame of original (i.e. incomplete) data}

\item{imp}{list of length \code{m} of imputed data sets}

\item{counts}{vector giving the name(s) of the count variables for which observed and imputed counts shall be compared}
}
\description{
Compare percentages of observed and imputed counts
}
\author{
Kristian Kleinke
}
