\name{countimp-package}
\alias{countimp-package}
\alias{countimp}
\docType{package}
\title{
Multiple Imputation of Incomplete Count Data}
\description{
The package provides functions for multiply imputing incomplete ordinary, overdispersed, zero-inflated and multilevel count data. The functions work as an add-on to the package \pkg{mice} (van Buuren and Groothuis-Oudshoorn, 2011). Multiple imputations are created within a sequential regressions MI framework by the \code{mice()} function.}
\details{
\tabular{ll}{
Package: \tab countimp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-15\cr
License: \tab GPL-3\cr
}
For an overview of how to use the package, including the most important functions, see 
\url{www.uni-bielefeld.de/soz/kds/pdf/countimp.pdf}.}
\author{
Kristian Kleinke and Jost Reinecke

Maintainer: Kristian Kleinke \url{mailto:kristian.kleinke@uni-bielefeld.de}

}
\references{
Kleinke, K., & Reinecke, J. (2013). \emph{countimp 1.0 -- A multiple imputation package for incomplete count data} [Technical Report]. University of Bielefeld, Faculty of Sociology, available from \url{www.uni-bielefeld.de/soz/kds/pdf/countimp.pdf}.

van Buuren, S., & Groothuis-Oudshoorn, K. (2011). MICE: Multivariate imputation by chained equations in R. \emph{Journal of Statistical Software}, 45(3), 1--67.
}
\keyword{ package }
\seealso{


}
\examples{

}
