% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crim4.R
\docType{data}
\name{crim4w}
\alias{crim4w}
\title{The CRIMOC data set in wide format}
\format{
A data frame with 2064 rows and 9 variables
}
\usage{
data(crim4w)
}
\description{
Dataset in wide format, see Reinecke and Weins (2013) for details.
}
\details{
\describe{
 \item{\code{id}}{participant identifier}
 \item{\code{FEMALE}}{Gender indicator}
 \item{\code{RE}}{school type dummy (intermediate branch)}
 \item{\code{GY}}{school type dummy (top level branch)}
 \item{\code{HY}}{school type dummy (lowest level)}
 \item{\code{ACRIM}}{Delinquency score -- time 1}
 \item{\code{BCRIM}}{Delinquency score -- time 2}
 \item{\code{CCRIM}}{Delinquency score -- time 3}
 \item{\code{DCRIM}}{Delinquency score -- time 4}
}
}
\references{
Reinecke, J., & Weins, C. (2013). The development of delinquency during adolescence: a comparison of missing data techniques. \emph{Quality & Quantity, 47(6)}, 3319--3334.
}
\keyword{datasets}
