% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.2l.pmm.R
\name{mice.impute.2l.pmm}
\alias{mice.impute.2l.pmm}
\title{Two-level predictive mean matching}
\usage{
mice.impute.2l.pmm(
  y,
  ry,
  x,
  type,
  intercept = TRUE,
  donors = 5,
  wy = NULL,
  ...
)
}
\arguments{
\item{y}{Numeric vector with incomplete data}

\item{ry}{Response pattern of \code{y} (\code{TRUE}=observed, \code{FALSE}=missing)}

\item{x}{matrix with \code{length(y)} rows containing complete covariates}

\item{type}{vector of length \code{ncol(x)} determining the imputation model; \code{type} is automatically extracted from the \code{predictorMatrix} argument of \code{mice()}.}

\item{intercept}{Logical. shall the intercept be included as a fixed and random effect?. \code{TRUE} = yes; \code{FALSE} = fixed effect only.}

\item{donors}{The size of the donor pool; default is 5.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value indicates locations in \code{y} for which imputations are created. Default is \code{!ry}}

\item{...}{additional arguments passed down from the main mice call}
}
\value{
vector with imputations
}
\description{
The function imputes an incomplete variable based on a normal linear mixed effects model. The model is estimated using function \code{glmmPQL()} from package \pkg{MASS}. Matching is done by \code{.pmm.match} from package \pkg{mice}.
}
\details{
Model specification / allowed entries in \code{mice}'s \code{predictorMatrix}:
   \itemize{
     \item 0 = variable not included in imputation model
     \item 1 = fixed effect
     \item 2 = fixex and random effect
     \item -2 = class variable
   }
}
\references{
\itemize{
\item Kleinke, K. (2016, September). Multiple Imputation of Multilevel Data by "Two-Level Predictive Mean Matching". Paper presented at the 50th Congress of the German Psychological Society (DGPs), Leipzig, Germany.
}
}
\author{
Kristian Kleinke
}
