% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countimp.R
\name{countimp}
\alias{countimp}
\title{\code{countimp} Wrapper function}
\usage{
countimp(...)
}
\arguments{
\item{...}{\code{mice()} arguments see \code{help("mice")}}
}
\description{
Depending on the entries of \code{mice}'s \code{method} argument, this function either calls \code{mice} directly or calls our variant to impute flat file zero-inflated count data.
}
\examples{
data(crim4w)
ini <- countimp(crim4w, maxit=0)
meth <- ini$method
meth[6:7] <- "hp"
meth[8:9] <- "pmm"
pred <- ini$predictorMatrix
pred[,"id"] <- 0
pred["ACRIM",] <- c(0,1,3,2,0,3,3,2,1)
imp <- countimp( data = crim4w, method = meth, predictorMatrix = pred)
imp$method
}
\author{
Kristian Kleinke
}
