% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crim4l.R
\docType{data}
\name{crim4l}
\alias{crim4l}
\title{The CRIMOC data set in long format}
\format{A data frame with 8256 rows and 6 variables}
\usage{
data(crim4l)
}
\description{
Dataset in long format, see Reinecke and Weins (2013) for details.
}
\details{
\describe{
 \item{\code{FEMALE}}{Gender indicator}
 \item{\code{RE}}{school type dummy (intermediate branch)}
 \item{\code{GY}}{school type dummy (top level branch)}
 \item{\code{TIME}}{TIME variable, indicating the respective panel wave (0,1,2,3); 0 is the first wave}
 \item{\code{DELINQ}}{Delinquency score in long format}
 \item{\code{ID}}{Participant identifier (here: cluster ID)}
}
}
\references{
Reinecke, J., & Weins, C. (2013). The development of delinquency during adolescence: a comparison of missing data techniques. \emph{Quality & Quantity, 47(6)}, 3319--3334.
}
\keyword{datasets}
